feat(concrete_core_wasm): automatize api generation
