+++ concrete-core-wasm/Cargo.toml
@@ -10,6 +10,7 @@ documentation = "https://docs.zama.ai/concrete/core-lib/"
 repository = "https://github.com/zama-ai/concrete-core"
 readme = "README.md"
 keywords = ["fully", "homomorphic", "encryption", "fhe", "cryptography"]
+build = "build/build.rs"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
@@ -19,15 +20,23 @@ wasm-bindgen = {version = "0.2.63", features = ["serde-serialize"]}
 js-sys = "0.3"
 console_error_panic_hook = "0.1.7"
 serde = { version = "1.0", features = ["derive"] }
+serde-wasm-bindgen = "0.4"
 paste = "1.0.7"
 
 [dependencies.concrete-core]
 path = "../concrete-core"
 default-features = false
-features = ["backend_default","backend_default_parallel", "backend_default_serialization"]
+features = ["backend_default", "backend_default_parallel", "backend_default_serialization", "backend_fft", "backend_fft_serialization"]
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3.13"
+
+[build-dependencies]
+concrete-core-representation = {path = "../concrete-core-representation"}
+syn={version="1.0", features=["full", "extra-traits"]}
+quote = "1.0"
+lazy_static = "1.4.0"
+
 
 #[profile.release]
 #lto = "fat"
