--- concrete-core-wasm/src/engines/mod.rs
@@ -1,20 +1,0 @@
-macro_rules! jserr {
-    ($expr: expr) => {
-        $expr.map_err(|e| wasm_bindgen::JsError::new(format!("{}", e).as_str()))
-    };
-}
-
-macro_rules! wrap {
-    ($newtype: ident, $expr: expr) => {
-        jserr!($expr).map($newtype)
-    };
-}
-
-mod default_engine;
-pub use default_engine::*;
-
-mod default_parallel_engine;
-pub use default_parallel_engine::*;
-
-mod default_serialization_engine;
-pub use default_serialization_engine::*;
